requirements.txt
Flask==3.1.2
mysql-connector-python==8.1.0
python-dotenv